:recycle: refactor: 전역화 필요한 타입 리팩토링
