+++ src/contexts/SocketProvider/useStomp/index.ts
@@ -1,5 +1,5 @@
-import { Notification } from "@contexts/AuthProvider/types";
 import { useAuthContext } from "@contexts/hooks";
+import { Notification } from "@domainTypes/.";
 import { CompatClient } from "@stomp/stompjs";
 import { useCallback, useEffect, useState } from "react";
 import { socketApi } from "service";
@@ -49,8 +49,9 @@ const useStomp: UseStomp = (token: string) => {
     console.log("SEND,Token", "destination:", destination, "body:", body);
 
     const bodyStringified = JSON.stringify(body);
-    if (compatClient && token)
+    if (compatClient && token) {
       compatClient.send(destination, { token }, bodyStringified);
+    }
   };
 
   return { isConnected, isLoading, sendAuth };
