+++ src/domainTypes/favorites.ts
@@ -1,0 +1,9 @@
+export interface Favorite {
+  favoriteId: number;
+  courtId: number;
+  courtName: string;
+  latitude: number;
+  longitude: number;
+  createdAt: string;
+  updatedAt: string;
+}
