+++ src/domainTypes/follows.ts
@@ -1,0 +1,5 @@
+export interface Follow {
+  followId: number;
+  followerId: number;
+  followingId: number;
+}
