+++ src/domainTypes/index.ts
@@ -1,0 +1,20 @@
+export { notificationTypes } from "./notifications";
+export { roles } from "./users";
+export { textures } from "./courts";
+export { status } from "./newCourts";
+
+export type { Notification, NotificationType } from "./notifications";
+export type { Follow } from "./follows";
+export type { Favorite } from "./favorites";
+export type { Reservation } from "./reservations";
+export type {
+  Role,
+  EditableUserProfile,
+  ProficiencyKeyUnion,
+  ProficiencyValueUnion,
+  PositionKeyUnion,
+  PositionValueUnion,
+} from "./users";
+export type { Coord } from "./map";
+export type { TextureValueUnion, TextureKeyUnion } from "./courts";
+export type { StatusValueUnion, StatusKeyUnion } from "./newCourts";
