+++ src/domainTypes/map.ts
